requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2-binary
gunicorn
marshmallow
requests
pytest-flask
flask-cors
